pv-prospect-data-versioner/pyproject.toml
[tool.poetry]
name = "pv-prospect-data-versioner"
version = "0.1.0"
description = ""
authors = ["RJ Ferrier <rjferrier@outlook.com>"]
readme = "README.md"
packages = [{include = "pv_prospect"}]

[project]
name = "pv-prospect-data-versioner"
version = "0.1.0"
description = ""
authors = [{name = "RJ Ferrier", email = "rjferrier@outlook.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc[gs]>=3,<4",
    "gitpython>=3,<4",
    "pv-prospect-common",
    "pv-prospect-etl",
]

[tool.poetry.dependencies]
python = "^3.12"
pv-prospect-common = {path = "../pv-prospect-common", develop = true}
pv-prospect-etl = {path = "../pv-prospect-etl", develop = true}
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
ruff = { version = "^0.13" }
mypy = { version = "^1.18" }
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core>=2.2.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
